+++ tests/Potion.Service.Tests/Remediation/RemediationTaskExecutorTests.cs
@@ -23,7 +23,8 @@ public class RemediationTaskExecutorTests
     {
         _validator.Setup(v => v.EnsureCommandIsAllowed(It.IsAny<string>()))
             .Returns((string c) => c);
-        _executor = new RemediationTaskExecutor(_logger.Object, _runner.Object, _validator.Object);
+        _executor = new RemediationTaskExecutor(_logger.Object, _runner.Object, _validator.Object,
+            new RemediationExecutionStats());
     }
 
     private static RemediationTaskDescriptor Descriptor(RemediationTaskOption? option = null) =>